split overview.go into mutliple files
